+++ src/test/java/flab/project/controller/DebatePostControllerTest.java
@@ -3,9 +3,12 @@ package flab.project.controller;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import flab.project.common.Constraints;
 import flab.project.config.baseresponse.ResponseEnum;
-import flab.project.config.baseresponse.SuccessResponse;
-import flab.project.data.dto.model.AddDebatePostRequest;
-import flab.project.template.PostFacadeTemplate;
+import flab.project.domain.feed.service.FanOutService;
+import flab.project.domain.post.enums.PostType;
+import flab.project.domain.post.model.AddDebatePostRequest;
+import flab.project.domain.post.controller.DebatePostController;
+import flab.project.domain.post.model.DebatePost;
+import flab.project.domain.post.template.PostFacadeTemplate;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -18,6 +21,7 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.ResultMatcher;
 
+import java.sql.Timestamp;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -34,7 +38,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@WebMvcTest(controllers = DebatePostController.class)
+@WebMvcTest(DebatePostController.class)
 class DebatePostControllerTest {
 
     private static final String ADD_DEBATE_POST_REQUEST_URL = "/posts/debate";
@@ -48,8 +52,10 @@ class DebatePostControllerTest {
     private ObjectMapper objectMapper;
     @MockBean
     private PostFacadeTemplate postFacadeTemplate;
+    @MockBean
+    private FanOutService fanOutService;
 
-    @WithMockUser
+    @WithMockUser(username = "1")
     @DisplayName("토론 게시물을 생성할 수 있다.")
     @Test
     void addDebatePost() throws Exception {
@@ -57,11 +63,27 @@ class DebatePostControllerTest {
         AddDebatePostRequest validDebatePostRequest = createAddDebatePost(validPostContent, validHashTagNames, validOptionContents);
 
         // when & then
+        DebatePost mockDebatePost = DebatePost.builder()
+                .postId(1L)
+                .userId(1L)
+                .content(validPostContent)
+                .hashTagNames(validHashTagNames)
+                .postType(PostType.DEBATE)
+                .likeCount(0)
+                .commentCount(0)
+                .createdAt(new Timestamp(System.currentTimeMillis()))
+                .isLike(false)
+                .isFollow(false)
+                .options(Set.of())
+                .selectedOptionId(null)
+                .build();
+
+        given(postFacadeTemplate.addPost(anyLong(), any(AddDebatePostRequest.class))).willReturn(mockDebatePost);
         validateAddDebatePostRequest(validDebatePostRequest, ADD_DEBATE_POST_REQUEST_URL, status().isOk(), SUCCESS);
         then(postFacadeTemplate).should().addPost(anyLong(), any(AddDebatePostRequest.class));
     }
 
-    @WithMockUser
+    @WithMockUser(username = "1")
     @DisplayName("토론 게시물을 생성할 때, 비어 있는 postContent가 있으면 INVALID_USER_INPUT을 반환한다.")
     @Test
     void addDebatePost_withEmptyPostContent() throws Exception {
@@ -73,7 +95,7 @@ class DebatePostControllerTest {
         validateAddDebatePostRequest(invalidDebatePostRequest, ADD_DEBATE_POST_REQUEST_URL, status().isBadRequest(), INVALID_USER_INPUT);
     }
 
-    @WithMockUser
+    @WithMockUser(username = "1")
     @DisplayName("토론 게시물을 생성할 때, 공백으로만 이루어진 postContent가 있으면 INVALID_USER_INPUT을 반환한다.")
     @Test
     void addDebatePost_withOnlyBlankPostContent() throws Exception {
@@ -85,7 +107,7 @@ class DebatePostControllerTest {
         validateAddDebatePostRequest(invalidDebatePostRequest, ADD_DEBATE_POST_REQUEST_URL, status().isBadRequest(), INVALID_USER_INPUT);
     }
 
-    @WithMockUser
+    @WithMockUser(username = "1")
     @DisplayName("토론 게시물을 생성할 때, 최대 길이를 초과한 postContent가 있으면 INVALID_USER_INPUT을 반환한다.")
     @Test
     void addDebatePost_withExceedMaxLengthOfPostContent() throws Exception {
@@ -97,7 +119,7 @@ class DebatePostControllerTest {
         validateAddDebatePostRequest(invalidDebatePostRequest, ADD_DEBATE_POST_REQUEST_URL, status().isBadRequest(), INVALID_USER_INPUT);
     }
 
-    @WithMockUser
+    @WithMockUser(username = "1")
     @DisplayName("토론 게시물을 생성할 때, 비어 있는 문자열의 hashTagName이 있으면 INVALID_USER_INPUT을 반환한다.")
     @Test
     void addDebatePost_withEmptyHashTagName() throws Exception {
@@ -109,7 +131,7 @@ class DebatePostControllerTest {
         validateAddDebatePostRequest(invalidDebatePostRequest, ADD_DEBATE_POST_REQUEST_URL, status().isBadRequest(), INVALID_USER_INPUT);
     }
 
-    @WithMockUser
+    @WithMockUser(username = "1")
     @DisplayName("토론 게시물을 생성할 때, 공백으로만 이루어진 hashTagName이 있으면 INVALID_USER_INPUT을 반환한다.")
     @Test
     void addDebatePost_withOnlyBlankHashTagName() throws Exception {
@@ -121,7 +143,7 @@ class DebatePostControllerTest {
         validateAddDebatePostRequest(invalidDebatePostRequest, ADD_DEBATE_POST_REQUEST_URL, status().isBadRequest(), INVALID_USER_INPUT);
     }
 
-    @WithMockUser
+    @WithMockUser(username = "1")
     @DisplayName("토론 게시물을 생성할 때, hashTagNames가 최대 개수를 초과하면 INVALID_USER_INPUT을 반환한다.")
     @Test
     void addDebatePost_withExceedMaxSizeOfHashTagNames() throws Exception {
@@ -134,7 +156,7 @@ class DebatePostControllerTest {
         validateAddDebatePostRequest(invalidDebatePostRequest, ADD_DEBATE_POST_REQUEST_URL, status().isBadRequest(), INVALID_USER_INPUT);
     }
 
-    @WithMockUser
+    @WithMockUser(username = "1")
     @DisplayName("토론 게시물을 생성할 때, hashTagNames 중 최대 길이를 초과한 hashTagName이 있으면 INVALID_USER_INPUT을 반환한다.")
     @Test
     void addDebatePost_withExceedMaxLengthOfHashTagName() throws Exception {
@@ -148,7 +170,7 @@ class DebatePostControllerTest {
         validateAddDebatePostRequest(invalidDebatePostRequest, ADD_DEBATE_POST_REQUEST_URL, status().isBadRequest(), INVALID_USER_INPUT);
     }
 
-    @WithMockUser
+    @WithMockUser(username = "1")
     @DisplayName("토론 게시물을 생성할 때, 비어 있는 문자열의 optionContent가 있으면 INVALID_USER_INPUT을 반환한다.")
     @Test
     void addDebatePost_withEmptyOptionContents() throws Exception {
@@ -160,7 +182,7 @@ class DebatePostControllerTest {
         validateAddDebatePostRequest(invalidDebatePostRequest, ADD_DEBATE_POST_REQUEST_URL, status().isBadRequest(), INVALID_USER_INPUT);
     }
 
-    @WithMockUser
+    @WithMockUser(username = "1")
     @DisplayName("토론 게시물을 생성할 때, 공백으로만 이루어진 optionContent가 있으면 INVALID_USER_INPUT을 반환한다.")
     @Test
     void addDebatePost_withOnlyBlankOptionContents() throws Exception {
@@ -172,7 +194,7 @@ class DebatePostControllerTest {
         validateAddDebatePostRequest(invalidDebatePostRequest, ADD_DEBATE_POST_REQUEST_URL, status().isBadRequest(), INVALID_USER_INPUT);
     }
 
-    @WithMockUser
+    @WithMockUser(username = "1")
     @DisplayName("토론 게시물을 생성할 때, optionContents 요소가 2가 아니면 INVALID_USER_INPUT을 반환한다.")
     @Test
     void addDebatePost_invalidSizeOfOptionContents() throws Exception {
@@ -193,7 +215,7 @@ class DebatePostControllerTest {
         validateAddDebatePostRequest(anotherInvalidDebatePostRequest, ADD_DEBATE_POST_REQUEST_URL, status().isBadRequest(), INVALID_USER_INPUT);
     }
 
-    @WithMockUser
+    @WithMockUser(username = "1")
     @DisplayName("토론 게시물을 생성할 때, optionContents 요소중 최대 길이를 초과한 요소가 있으면 INVALID_USER_INPUT을 반환한다.")
     @Test
     void addDebatePost_exceedMaxLengtheOfOptionContents() throws Exception {
